PIM-3595: move orm and mongo cursors in a dedicated folder
